Exam created from older versions of artemis cannot be deleted if an exercise group has been deleted. This is due to the ordered column of the exercise groups return null values

+++ src/main/java/de/tum/in/www1/artemis/web/rest/ExamResource.java
@@ -336,8 +336,10 @@ public class ExamResource {
         auditEventRepository.add(auditEvent);
         log.info("User " + user.getLogin() + " has requested to delete the exam {}", exam.getTitle());
         for (ExerciseGroup exerciseGroup : exam.getExerciseGroups()) {
-            for (Exercise exercise : exerciseGroup.getExercises()) {
-                exerciseService.delete(exercise.getId(), false, false);
+            if (exerciseGroup != null) {
+                for (Exercise exercise : exerciseGroup.getExercises()) {
+                    exerciseService.delete(exercise.getId(), false, false);
+                }
             }
         }
